requirements.txt
faiss-cpu>=1.8.0
networkx>=3.4.0
numpy>=1.26.0
openai>=1.30.0
sentence-transformers>=2.7.0
